+++ cmd/thanos/config.go
@@ -182,6 +182,7 @@ type queryConfig struct {
 	dnsSDInterval time.Duration
 	httpMethod    string
 	dnsSDResolver string
+	step          time.Duration
 }
 
 func (qc *queryConfig) registerFlag(cmd extkingpin.FlagClause) *queryConfig {
@@ -198,6 +199,8 @@ func (qc *queryConfig) registerFlag(cmd extkingpin.FlagClause) *queryConfig {
 		Default("POST").EnumVar(&qc.httpMethod, "GET", "POST")
 	cmd.Flag("query.sd-dns-resolver", "Resolver to use. Possible options: [golang, miekgdns]").
 		Default("golang").Hidden().StringVar(&qc.dnsSDResolver)
+	cmd.Flag("query.default-step", "Default range query step to use. This is only used in stateless Ruler and alert state restoration.").
+		Default("1s").DurationVar(&qc.step)
 	return qc
 }